Q1 진행

+++ EmptyViewer/Light.h
@@ -9,5 +9,5 @@ public:
 
 	vec3 pos;
 
-	int illumination;
+	float illumination;
 };
